novels/backend/requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0

# Database
psycopg2-binary>=2.9.9

# AI Services
openai>=1.12.0
python-dotenv>=1.0.0

# Audio Processing
pydub>=0.25.1
ffmpeg-python>=0.2.0

# Object Storage (MinIO)
minio>=7.2.0

# Utils
tenacity>=8.2.3
